+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 ### Added
+- `clustering:` Ported `VectorWithNorm` datatype
 - `linalg:` Ported `DenseVector` and `SparseVector` datatypes, plus their `Vectors` factory
 
 <small>(formatted per [keepachangelog-1.1.0](http://keepachangelog.com/en/1.0.0/))</small>